Fix 73 - Fix fetching user's playlists
I think I've found something: I passed `$acb` as the callback for the `_call` method, which would return the result and the response object as the second argument. But `$acb` would consider a second argument an error. Thus would cut short and return only the first page of results, assuming the others failed.

By explicitly calling `$acb->(shift)` we only pass the result set. Which allows `achain` to evaluate all result sets of all the calls.